requirements.txt
anyio==3.7.1
boto3==1.34.91
botocore==1.34.91
brevitas==0.8.0
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
coloredlogs==15.0.1
-e git+https://github.com/mtulla/concrete-ml.git@79b1719a896fca2aea430bf98da2865f6bd63bbb#egg=concrete_ml&subdirectory=../../concrete-ml
concrete-python==2.6.0rc1
dependencies==2.0.1
dill==0.3.8
exceptiongroup==1.2.1
fastapi==0.103.2
filelock==3.13.4
flatbuffers==24.3.25
fsspec==2024.3.1
gitdb==4.0.11
GitPython==3.1.41
h11==0.14.0
huggingface-hub==0.22.2
humanfriendly==10.0
hummingbird-ml==0.4.8
idna==3.7
importlib_resources==6.4.0
iniconfig==2.0.0
jmespath==1.0.1
joblib==1.4.0
jsonpickle==3.0.4
-e git+https://github.com/mtulla/minGPT.git@f7f38be6ae573a1485b49a8b3b7444f3a21eba79#egg=minGPT
mpmath==1.3.0
networkx==3.2.1
numpy==1.23.5
onnx==1.15.0
onnxconverter-common==1.13.0
onnxmltools==1.11.0
onnxoptimizer==0.3.13
onnxruntime==1.17.3
packaging==24.0
pandas==2.0.3
pluggy==1.5.0
protobuf==3.20.3
psutil==5.9.8
pydantic==1.10.15
pytest==8.1.1
pytest-json-report==1.5.0
pytest-metadata==3.1.1
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
regex==2024.4.16
requests==2.31.0
s3transfer==0.10.1
safetensors==0.4.3
scikit-learn==1.1.3
scipy==1.10.1
six==1.16.0
skl2onnx==1.12
skops==0.5.0
skorch==0.11.0
smmap==5.0.1
sniffio==1.3.1
starlette==0.27.0
sympy==1.12
tabulate==0.9.0
threadpoolctl==3.4.0
tokenizers==0.19.1
tomli==2.0.1
torch==1.13.1
tqdm==4.66.2
transformers==4.40.1
typing_extensions==4.5.0
tzdata==2024.1
urllib3==1.26.18
uvicorn==0.21.1
xgboost==1.6.2
z3-solver==4.13.0.0
zipp==3.18.1
